+++ lib/api-client/index.ts
@@ -2,9 +2,7 @@
 /* istanbul ignore file */
 /* tslint:disable */
 /* eslint-disable */
-import './setup';
 export { ApiError } from './core/ApiError';
-
 export { CancelablePromise, CancelError } from './core/CancelablePromise';
 export { OpenAPI } from './core/OpenAPI';
 export type { OpenAPIConfig } from './core/OpenAPI';
@@ -14,6 +12,9 @@ export type { AuthTokens } from './models/AuthTokens';
 export type { Cart } from './models/Cart';
 export type { CartItem } from './models/CartItem';
 export type { CartItemQuantity } from './models/CartItemQuantity';
+export type { CartSession } from './models/CartSession';
+export type { CatalogFacets } from './models/CatalogFacets';
+export type { Category } from './models/Category';
 export type { CheckoutSession } from './models/CheckoutSession';
 export type { Currency } from './models/Currency';
 export type { Cursor } from './models/Cursor';
@@ -25,8 +26,10 @@ export type { Money } from './models/Money';
 export type { Order } from './models/Order';
 export type { OrderCancellation } from './models/OrderCancellation';
 export type { OrderConfirmation } from './models/OrderConfirmation';
+export type { OrderLineItem } from './models/OrderLineItem';
 export { OrderStatus } from './models/OrderStatus';
 export type { OrderSummary } from './models/OrderSummary';
+export type { OrderTotals } from './models/OrderTotals';
 export type { Payment } from './models/Payment';
 export type { PaymentIntent } from './models/PaymentIntent';
 export { PaymentStatus } from './models/PaymentStatus';