+++ next.config.ts
@@ -2,6 +2,9 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   output: "standalone",
+  logging: {
+    fetches: { fullUrl: true },
+  },
 };
 
 export default nextConfig;